+++ tests/unit/test_main.py
@@ -25,6 +25,7 @@ def test__parse_args__from():
         strict=False,
         path=None,
         pattern_prefix=None,
+        ignore_untracked=False,
     )
     assert parse_args(["from", "git"]).vcs == "git"
     assert parse_args(["from", "git", "--tag-branch", "foo"]).tag_branch == "foo"
@@ -53,6 +54,7 @@ def test__parse_args__from():
     assert parse_args(["from", "any", "--strict"]).strict is True
     assert parse_args(["from", "any", "--path", "/tmp"]).path == "/tmp"
     assert parse_args(["from", "any", "--pattern-prefix", "foo-"]).pattern_prefix == "foo-"
+    assert parse_args(["from", "any", "--ignore-untracked"]).ignore_untracked is True
     assert parse_args(["from", "subversion", "--tag-dir", "foo"]).tag_dir == "foo"
 
     with pytest.raises(SystemExit):